Removed Sticky Header
Removed Sticky Header due to bug when closing keyboard, instead used appbar : AppBar( flexiblespace: ).
More layout changes.
Added Hero animation.
