--- lib/inputtheme.dart
@@ -1,22 +1,0 @@
-// import 'package:flutter/material.dart';
-// import 'constants.dart';
-
-// class InputTheme {
-//   final Color bordercolor = const Color.fromRGBO(10, 10, 10, 1);
-
-//   TextStyle _buildTextStyle(Color? color, {double size = 16.0}) {
-//     return TextStyle(
-//       color: color,
-//       fontSize: size,
-//     );
-//   }
-
-//   InputDecorationTheme theme() => const InputDecorationTheme(
-//         contentPadding: EdgeInsets.all(16),
-//         isDense: true,
-//         floatingLabelBehavior: FloatingLabelBehavior.auto,
-//         constraints: BoxConstraints(maxWidth: 200),
-
-//         suffixStyle: _buildTextStyle(Colors.black),
-//       );
-// }
