+++ pubspec.yaml
@@ -32,7 +32,6 @@ dependencies:
   flutter:
     sdk: flutter
 
-  sticky_headers: "^0.3.0"
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
